requirements.txt
streamlit
rembg
Pillow
streamlit-extras